+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+### Added
+- #463 Setting the display name of the Launcher at runtime
+
 ## [v1.5.0.0]
 ### Added
 - #1214 Additional log entries